+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -10,7 +10,7 @@ public class HanoiRhymer extends DefaultCountingOutRhymer {
 
     @Override
     public void countIn(int in) {
-        if (!callCheck() && in > peekABoo())
+        if (!callCheck() && in > countOut())
             totalRejected++;
         else
             super.countIn(in);
