+++ src/main/java/edu/kis/vh/nursery/IntArrayStack.java
@@ -1,0 +1,34 @@
+package edu.kis.vh.nursery;
+
+public class IntArrayStack {
+    private final int n = 12;
+    private final int[] numbers = new int[n];
+
+    public int total = -1;
+
+    public void countIn(int in) {
+        if (!isFull())
+            numbers[++total] = in;
+    }
+
+    public boolean callCheck() {
+        return total == -1;
+    }
+
+    public boolean isFull() {
+        return total == n-1;
+    }
+
+    protected int peekABoo() {
+        if (callCheck())
+            return -1;
+        return numbers[total];
+    }
+
+
+    public int countOut() {
+        if (callCheck())
+            return -1;
+        return numbers[total--];
+    }
+}
